Surface H3 SETTINGS/GOAWAY as events and add H3 close() (#110)
* Surface H3 SETTINGS/GOAWAY as events and add H3 close()

The HTTP/3 read path now pushes a Settings event when the peer's control
stream delivers its SETTINGS frame and a Goaway event when a GOAWAY arrives,
matching the HTTP/2 model where these reach the integrator through
next_event() rather than poll-only getters. The shared SettingPair value and
Goaway last_stream_id widen to u64 so HTTP/3's 62-bit settings values and
stream ids are lossless; HTTP/2's narrower values fit unchanged.

H3Connection.close(error_code=H3_NO_ERROR, reason=b"") sends a caller-chosen
application CONNECTION_CLOSE, mirroring H2Connection.close().

* Reject an H3 close() error code past the 62-bit QUIC range

Without the bound, conn.close(1 << 62) reached the varint encoder, which
failed and surfaced as a misleading MemoryError. Validate the code in the
adapter like the stream reset path does, raising ValueError.

+++ src/core/events.zig
@@ -64,7 +64,7 @@ pub const EndOfMessage = struct {
 /// (and knows an ACK is owed). HTTP/2 only.
 pub const SettingPair = struct {
     id: u16,
-    value: u32,
+    value: u64,
 };
 
 /// A stream was reset by the peer (RFC 9113 6.4). HTTP/2 only.
@@ -81,9 +81,11 @@ pub const StreamReset = struct {
     error_code: u64,
 };
 
-/// The peer is shutting the connection down (RFC 9113 6.8). HTTP/2 only.
+/// The peer is shutting the connection down: HTTP/2 GOAWAY (RFC 9113 6.8) or
+/// HTTP/3 GOAWAY (RFC 9114 5.2). `last_stream_id` is u64 to hold the HTTP/3
+/// 62-bit id losslessly; HTTP/2's 31-bit id fits the same field.
 pub const Goaway = struct {
-    last_stream_id: u32,
+    last_stream_id: u64,
     error_code: u32,
     debug: []const u8 = &.{},
 };
